refactor: separate hook into new file
